fixed incorrect type for the ghl estimate's discount value (due to the api docs reporting it as an int when it can be a float)

+++ src/api_framework/models/gohighlevel/estimates.py
@@ -112,7 +112,7 @@ class EstimateResponse(BaseModel):
     )
     # discount: Mapping[str, Any]
     discount_type: str
-    discount_value: int
+    discount_value: float
     title: str
     name: str
     items: tuple[EstimateItemsResponse, ...]
@@ -223,7 +223,7 @@ class EstimateTemplateResponse(BaseModel):
         validation_alias = "deleted"
     )
     discount_type: str
-    discount_value: int
+    discount_value: float
     title: str
     name: str
     items: tuple[EstimateItemsResponse, ...]
